Выполняет модуляцию проекта
